Fix: Id data type issue resolved.

+++ src/main/java/com/smart/event/event_management_system/dto/BookingRequestDto.java
@@ -2,7 +2,7 @@ package com.smart.event.event_management_system.dto;
 
 public class BookingRequestDto {
     private Long userId;
-    private Integer eventId;
+    private Long eventId;
     private int numberOfTickets;
     
     public Long getUserId() {
@@ -11,10 +11,10 @@ public class BookingRequestDto {
     public void setUserId(Long userId) {
         this.userId = userId;
     }
-    public Integer getEventId() {
+    public Long getEventId() {
         return eventId;
     }
-    public void setEventId(Integer eventId) {
+    public void setEventId(Long eventId) {
         this.eventId = eventId;
     }
     public int getNumberOfTickets() {
