+++ src/main/java/com/smart/event/event_management_system/dto/PaymentRequestDto.java
@@ -1,14 +1,14 @@
 package com.smart.event.event_management_system.dto;
 
 public class PaymentRequestDto {
-    private Integer bookingId;
+    private Long bookingId;
     private String mode;
 
     
-    public Integer getBookingId() {
+    public Long getBookingId() {
         return bookingId;
     }
-    public void setBookingId(Integer bookingId) {
+    public void setBookingId(Long bookingId) {
         this.bookingId = bookingId;
     }
     public String getMode() {
